test: get `cargo test --tests` working
Broken tests are disabled
