+++ azure-pipelines.yml
@@ -52,7 +52,7 @@ jobs:
       - tokio-sync
       # - tokio-threadpool
       # - tokio-timer
-      # - tokio-test
+      - tokio-test
       # - tokio-trace
       # - tokio-trace/tokio-trace-core
       # - tokio-trace/test-log-support
