+++ tokio-test/src/macros.rs
@@ -112,44 +112,3 @@ macro_rules! assert_elapsed {
     };
 }
 */
-
-#[cfg(test)]
-mod tests {
-    use futures::{future, Async, Future, Poll};
-
-    #[test]
-    fn assert_ready() {
-        let mut fut = future::ok::<(), ()>(());
-        assert_ready!(fut.poll());
-        let mut fut = future::ok::<(), ()>(());
-        assert_ready!(fut.poll(), "some message");
-    }
-
-    #[test]
-    #[should_panic]
-    fn assert_ready_err() {
-        let mut fut = future::err::<(), ()>(());
-        assert_ready!(fut.poll());
-    }
-
-    #[test]
-    fn assert_not_ready() {
-        let poll: Poll<(), ()> = Ok(Async::NotReady);
-        assert_not_ready!(poll);
-        assert_not_ready!(poll, "some message");
-    }
-
-    #[test]
-    #[should_panic]
-    fn assert_not_ready_err() {
-        let mut fut = future::err::<(), ()>(());
-        assert_not_ready!(fut.poll());
-    }
-
-    #[test]
-    fn assert_ready_eq() {
-        let mut fut = future::ok::<(), ()>(());
-        assert_ready_eq!(fut.poll(), ());
-    }
-
-}
