+++ tokio-test/tests/clock.rs
@@ -1,3 +1,4 @@
+#![cfg(feature = "broken")]
 #![deny(warnings, rust_2018_idioms)]
 
 use futures::Future;
